requirements.txt
onnxruntime
numpy
nibabel
SimpleITK
scikit-image
Pillow
flask
reportlab
pydicom
